separate concerns better
