+++ shared/actions/users-gen.tsx
@@ -6,12 +6,14 @@ import * as RPCTypes from '../constants/types/rpc-gen'
 export const resetStore = 'common:resetStore' // not a part of users but is handled by every reducer. NEVER dispatch this
 export const typePrefix = 'users:'
 export const blockUser = 'users:blockUser'
+export const getBio = 'users:getBio'
 export const updateBio = 'users:updateBio'
 export const updateBrokenState = 'users:updateBrokenState'
 export const updateFullnames = 'users:updateFullnames'
 
 // Payload Types
 type _BlockUserPayload = {readonly username: string}
+type _GetBioPayload = {readonly username: string}
 type _UpdateBioPayload = {readonly userCard: RPCTypes.UserCard; readonly username: string}
 type _UpdateBrokenStatePayload = {readonly newlyBroken: Array<string>; readonly newlyFixed: Array<string>}
 type _UpdateFullnamesPayload = {readonly usernameToFullname: {[username: string]: string}}
@@ -22,6 +24,7 @@ type _UpdateFullnamesPayload = {readonly usernameToFullname: {[username: string]
  */
 export const createUpdateBio = (payload: _UpdateBioPayload): UpdateBioPayload => ({payload, type: updateBio})
 export const createBlockUser = (payload: _BlockUserPayload): BlockUserPayload => ({payload, type: blockUser})
+export const createGetBio = (payload: _GetBioPayload): GetBioPayload => ({payload, type: getBio})
 export const createUpdateBrokenState = (payload: _UpdateBrokenStatePayload): UpdateBrokenStatePayload => ({
   payload,
   type: updateBrokenState,
@@ -33,6 +36,7 @@ export const createUpdateFullnames = (payload: _UpdateFullnamesPayload): UpdateF
 
 // Action Payloads
 export type BlockUserPayload = {readonly payload: _BlockUserPayload; readonly type: typeof blockUser}
+export type GetBioPayload = {readonly payload: _GetBioPayload; readonly type: typeof getBio}
 export type UpdateBioPayload = {readonly payload: _UpdateBioPayload; readonly type: typeof updateBio}
 export type UpdateBrokenStatePayload = {
   readonly payload: _UpdateBrokenStatePayload
@@ -47,6 +51,7 @@ export type UpdateFullnamesPayload = {
 // prettier-ignore
 export type Actions =
   | BlockUserPayload
+  | GetBioPayload
   | UpdateBioPayload
   | UpdateBrokenStatePayload
   | UpdateFullnamesPayload
