+++ shared/actions/users.tsx
@@ -3,9 +3,7 @@ import * as Saga from '../util/saga'
 import * as Container from '../util/container'
 import * as EngineGen from './engine-gen-gen'
 import * as UsersGen from './users-gen'
-import * as Chat2Gen from './chat2-gen'
 import * as RPCTypes from '../constants/types/rpc-gen'
-import * as Chat2Constants from '../constants/chat2'
 import {TypedState} from '../util/container'
 
 const onIdentifyUpdate = (
@@ -17,33 +15,25 @@ const onIdentifyUpdate = (
     newlyFixed: action.payload.params.okUsernames || [],
   })
 
-const fetchUserBio = async (state: TypedState, action: Chat2Gen.SelectConversationPayload) => {
-  const {conversationIDKey} = action.payload
-  const meta = Chat2Constants.getMeta(state, conversationIDKey)
-  const otherParticipants = Chat2Constants.getRowParticipants(meta, state.config.username || '')
-  if (otherParticipants.count() === 1) {
-    // we're in a one-on-one convo
-    const username = otherParticipants.first('')
-    if (username === '') {
-      return // if for some reason we get a garbage username, don't do anything
-    }
-    if (state.users.infoMap.get(username, {bio: undefined}).bio) {
-      return // don't re-fetch bio if we already have one cached
-    }
+// shouldn't know anything about chat stuff, only username => rpc call
+const getBio = async (state: TypedState, action: UsersGen.GetBioPayload) => {
+  const {username} = action.payload
 
-    const userCard = await RPCTypes.userUserCardRpcPromise({useSession: true, username})
-    if (!userCard) {
-      return // don't do anything if we don't get a good response from rpc
-    }
-
-    return UsersGen.createUpdateBio({userCard, username}) // set bio in user infomap
+  if (state.users.infoMap.get(username, {bio: undefined}).bio) {
+    return // don't re-fetch bio if we already have one cached
   }
-  return
+
+  const userCard = await RPCTypes.userUserCardRpcPromise({useSession: true, username})
+  if (!userCard) {
+    return // don't do anything if we don't get a good response from rpc
+  }
+
+  return UsersGen.createUpdateBio({userCard, username}) // set bio in user infomap
 }
 
 function* usersSaga() {
   yield* Saga.chainAction2(EngineGen.keybase1NotifyUsersIdentifyUpdate, onIdentifyUpdate, 'onIdentifyUpdate')
-  yield* Saga.chainAction2(Chat2Gen.selectConversation, fetchUserBio, 'fetchUpdateBio')
+  yield* Saga.chainAction2(UsersGen.getBio, getBio)
 }
 
 export default usersSaga
